Move custom nodes into nodes/ package
