+++ integration_tests/README.md
@@ -3,7 +3,7 @@
 These tests execute small ComfyUI workflows headlessly. To resolve built-in nodes,
 clone the ComfyUI repository into `ComfyUI_repo/` at the project root or otherwise
 ensure it is discoverable on `PYTHONPATH`.
-`conftest.py` automatically prepends this path and the `custom_nodes/` folder when
+`conftest.py` automatically prepends this path and the `nodes/` folder when
 running the tests.
 
 Run all integration tests:
