+++ integration_tests/conftest.py
@@ -4,6 +4,6 @@ import pathlib
 ROOT = pathlib.Path(__file__).parent.parent.resolve()
 comfy_src = ROOT / "ComfyUI_repo"
 sys.path.insert(0, str(comfy_src))
-sys.path.insert(0, str(ROOT / "custom_nodes"))
+sys.path.insert(0, str(ROOT / "nodes"))
 
 pytest_plugins = ["tests.conftest"]
