+++ nodes/conditional_save_image.py
