+++ nodes/string_utils.py
